+++ packages/studio-base/src/i18n/en/general.ts
