+++ packages/studio-base/src/i18n/en/index.ts
