+++ packages/studio-base/src/i18n/en/preferences.ts
