+++ packages/studio-base/src/i18n/index.ts
@@ -2,6 +2,10 @@
 // License, v2.0. If a copy of the MPL was not distributed with this
 // file, You can obtain one at http://mozilla.org/MPL/2.0/
 
+import i18n from "i18next";
+import LanguageDetector from "i18next-browser-languagedetector";
+import { initReactI18next } from "react-i18next";
+
 import { enPreferences, enGeneral } from "./en";
 import { zhPreferences, zhGeneral } from "./zh";
 
@@ -17,3 +21,20 @@ export const translations = {
 };
 
 export type Locale = keyof typeof translations;
+
+export const defaultNS = "general";
+
+export async function initI18n(): Promise<void> {
+  await i18n
+    .use(LanguageDetector)
+    .use(initReactI18next)
+    .init({
+      resources: translations,
+      detection: {
+        order: ["localStorage", "navigator"],
+        caches: ["localStorage"],
+      },
+      fallbackLng: "en",
+      defaultNS,
+    });
+}
