+++ packages/studio-base/src/i18n/zh/general.ts
