+++ packages/studio-base/src/i18n/zh/index.ts
