+++ packages/studio-base/src/i18n/zh/preferences.ts
