+++ packages/studio-base/src/index.ts
@@ -30,7 +30,7 @@ export type { IDataSourceFactory } from "./context/PlayerSelectionContext";
 export { default as installDevtoolsFormatters } from "./util/installDevtoolsFormatters";
 export { default as overwriteFetch } from "./util/overwriteFetch";
 export { default as waitForFonts } from "./util/waitForFonts";
-export { initI18n } from "./util/initI18n";
+export { initI18n } from "./i18n";
 export { IdbExtensionLoader } from "./services/IdbExtensionLoader";
 export type { ExtensionLoader } from "./services/ExtensionLoader";
 export type { ExtensionInfo, ExtensionNamespace } from "./types/Extensions";
