+++ packages/studio-base/src/typings/i18next.d.ts
@@ -2,8 +2,7 @@
 // License, v2.0. If a copy of the MPL was not distributed with this
 // file, You can obtain one at http://mozilla.org/MPL/2.0/
 
-import { translations } from "@foxglove/studio-base/locales";
-import { defaultNS } from "@foxglove/studio-base/util/initI18n";
+import { translations, defaultNS } from "@foxglove/studio-base/i18n";
 
 declare module "i18next" {
   interface CustomTypeOptions {
